Moved files from src/ to src/pages
