+++ src/pages/index.ts
@@ -2,13 +2,12 @@ import {
   toggleVisibilityTemporarily,
   setDisplayFlex,
   setDisplayNone
-} from './utils/utils'
+} from '../utils/utils'
 
 function index(): void {
   // Ejercicio: Clase 13/09/2021
   // Guardamos el nodo donde voy a extraer el dato
   saveDataAnchor()
-
   function saveDataAnchor(): void {
     const anchor: HTMLElement | null = document.getElementById('a');
     // Si el elemento con id = 'a' existe:
@@ -82,27 +81,25 @@ function index(): void {
 
 
 
-
   // Ejercicio: Clase 14/09/2021 (1/3)
   // Validación de entrada de datos en un formulario
-  const newsLetter = document.getElementById('news-letter') as HTMLElement
-  const successfulMessage = document.getElementById('successful-message') as HTMLElement
-  const invalidMessage = document.getElementById('invalid-message') as HTMLElement
-
-  const emailsBanned: string[] = [
-    'aquiles_bailo@yahoo.com',
-    'susana_oria@hotmail.com',
-    'aquiles_bailo@hotmail.com',
-    'armando_estebanquito@gmail.com',
-    'example@gmail.com',
-    'example@yahoo.com',
-    'example@hotmail.com',
-    'malapalabra@gmail.com'
-  ]
-
   addListenerForm()
-
   function addListenerForm (): void {
+    const newsLetter = document.getElementById('news-letter') as HTMLElement
+    const invalidMessage = document.getElementById('invalid-message') as HTMLElement
+    const successfulMessage = document.getElementById('successful-message') as HTMLElement
+
+    const emailsBanned: string[] = [
+      'aquiles_bailo@yahoo.com',
+      'susana_oria@hotmail.com',
+      'aquiles_bailo@hotmail.com',
+      'armando_estebanquito@gmail.com',
+      'example@gmail.com',
+      'example@yahoo.com',
+      'example@hotmail.com',
+      'malapalabra@gmail.com'
+    ]
+
     newsLetter.addEventListener('submit', function(e: Event): void {
       e.preventDefault()
       console.log('Dentro del boton')
