+++ tsconfig.json
@@ -28,5 +28,5 @@
     "skipLibCheck": true                                 /* Skip type checking all .d.ts files. */
   },
   "exclude": ["node_modules"],
-  "include": ["src/index.ts", "src/productlist.ts", "src/productdetails.ts"]
+  "include": ["src/pages/index.ts", "src/pages/productlist.ts", "src/pages/productdetails.ts"]
 }
